requirements.txt
requests>=2.28.0
pandas>=2.1.0
python-dotenv>=1.0.0
setuptools>=65.0.0
folium>=0.14.0
